requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
Django==5.1.5
django-cors-headers==4.7.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
exceptiongroup==1.2.2
idna==3.10
iniconfig==2.0.0
Markdown==3.7
packaging==24.2
pluggy==1.5.0
PyJWT==2.10.1
pytest==8.3.4
requests==2.32.3
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
